+++ plugins/sidepad/hooks/ask/index.ts
@@ -1,3 +1,4 @@
 export * as default from '.';
+export * from './ask-room-of';
 export * from './ask-text-of';
 export * from './fitted-ask-text-of';